Refactor chart toggle logic to be closer to usage
